% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/psychometricAnt.R
\name{psychometricAnt}
\alias{psychometricAnt}
\title{Psychometric Antonymns}
\usage{
psychometricAnt(Data1)
}
\arguments{
\item{Data1}{This is a matrix of person by item responses to polytomous items.}
}
\description{
TGenerates scores for psychometric antonyms, a measure of
within-person consistency.
}
\examples{
person1<-c(1,1,1,1,1,1,1,1,1)
person2<-c(1,2,3,3,2,1,2,2,2)
values<-rbind(person1,person2)
psychometricAnt(values)
}
\keyword{cleaning}
\keyword{data}

