% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/psychometricSyn.R
\name{psychometricSyn}
\alias{psychometricSyn}
\title{Psychometric Synonyms}
\usage{
psychometricSyn(Data1)
}
\arguments{
\item{Data1}{This is a matrix of person by item responses to polytomous items.}
}
\description{
Generates scores for psychometric synonyms, a measure of
within-person consistency.
}
\examples{
person1<-c(1,1,1,1,1,1,1,1,1)
person2<-c(1,2,3,3,2,1,2,2,2)
values<-rbind(person1,person2)
psychometricSyn(values)
}
\keyword{cleaning}
\keyword{data}

